1. Created validation for input text to ensure 3 characters are selected, 2. named generated SVG file, 3. Print messages on the console, 4. Moved files and structured properly
