+++ lib/util/colorList.js
