+++ lib/util/console-colors.js
