+++ lib/util/createFileSVG.js
