+++ lib/util/generateLogo.js
@@ -1,4 +1,4 @@
-const Shapes = require("./shapes.js");
+const Shapes = require("../shapes");
 const selectHex = require("./selectHexNumber.js");
 const createSVG = require("./createFileSVG.js");
 
@@ -19,7 +19,7 @@ function generateLogo(answers) {
 
   const templateSVG = createSVG({ shapeSVG, hexText, text });
 
-  generateLogo.prototype.something = makeShape;
+  generateLogo.prototype.shape = makeShape;
   return templateSVG;
 }
