+++ lib/util/renderColorList.js
@@ -1,4 +1,4 @@
-const colors = require("./colorList");
+const colors = require("./colorList.js");
 
 function renderColorList() {
   let colorList = [];
