+++ lib/util/selectHexNumber.js
@@ -1,4 +1,4 @@
-const colors = require("./colorList");
+const colors = require("./colorList.js");
 
 const selectHexNumber = function (selectColor) {
   for (const key in colors) {
